ref(internal): packages.js => projects.js
